requirements.txt
pkg-resources==0.0.0
psycopg2==2.8.6
speedtest-cli==2.1.2
